Split subscription testing out
