+++ monitoring/loadtest/locust_files/conf.py
@@ -1,0 +1,4 @@
+# OAUTH_HOST = "http://dummy-oauth.dss-main.svc:8085/token"
+# DSS_ENDPOINT = "http://http-gateway.dss-main.svc:8080/v1/dss"
+OAUTH_HOST = "http://34.123.61.246/token"
+DSS_ENDPOINT = "https://pcharlie1.interussplatform.dev/v1/dss"
